集成Mybatis官方代码生成器,访问数据库成功

+++ member/src/main/java/com/example/train/member/controller/MemberController.java
@@ -14,5 +14,5 @@ public class MemberController {
     private MemberService memberService;
 
     @GetMapping("/count")
-    public int count(){return memberService.count();}
+    public Long count(){return memberService.count();}
 }
